Testing White Noise in Time Series

The function Wnoise.test computes the test statistic for white noise in time series based on the variance scale exponent. The null hypothesis is that the time series is independent white noise, while the alternative hypothesis is that the time series is a non-independent stochastic process.

Usage

Wnoise.test(x, m = 0.5, n = NULL)

Arguments

x

A time series vector.

m

A parameter to control the number of scales. Default is 0.5.

n

The number of scales. If NULL, it will be calculated as floor(N^m).

Value

A list with class "Wnoise.test" containing the following components:

Wnoise

the test statistic

df

the degrees of freedom of the test.

p.value

the p-value of the test.

References

Fu, H., Chen, W., & He, X.-J. (2018). On a class of estimation and test for long memory. In Physica A: Statistical Mechanics and its Applications (Vol. 509, pp. 906–920). Elsevier BV. https://doi.org/10.1016/j.physa.2018.06.092

Examples

## Test white noise in time series
library(pracma)

set.seed(123)
data("brown72")
x72 <- brown72                          #  H = 0.72
xgn <- rnorm(1024)                      #  H = 0.50
xlm <- numeric(1024); xlm[1] <- 0.1     #  H = 0.43
for (i in 2:1024) xlm[i] <- 4 * xlm[i-1] * (1 - xlm[i-1])

Wnoise.test(x72)
#> Wnoise Test
#> 
#> Wnoise statistic: 135.1091 
#> degrees of freedom: 31 
#> p-value: 5.884182e-15 
#> 
#> alternative hypothesis: non-independent stochastic process
Wnoise.test(xgn)
#> Wnoise Test
#> 
#> Wnoise statistic: 27.9873 
#> degrees of freedom: 31 
#> p-value: 0.3781797 
#> 
#> alternative hypothesis: non-independent stochastic process
Wnoise.test(xlm)
#> Wnoise Test
#> 
#> Wnoise statistic: 19.58707 
#> degrees of freedom: 31 
#> p-value: 0.05575128 
#> 
#> alternative hypothesis: non-independent stochastic process